Client가 recv 한 후 m_vRecv_Queue에 데이터를 넣고, 매 렌더마다 Queue의 사이즈를 확인하여 데이터가 존재하면 카메라 위치를 이동.

+++ Server/KSLFS_Server/Server.h
@@ -15,9 +15,9 @@ enum SendType
 
 struct SendPosition {
 	SendType type;
-	int x;
-	int y;
-	int z;
+	float x;
+	float y;
+	float z;
 };
 
 class CServer
